Reconstruct client side code
