+++ portal-app/src/App.js
@@ -3,14 +3,14 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 
 import "./App.css";
 import "./index.css";
-import Home from "./pages/Home";
+import Home from "./pages/home";
 import { UploadContent } from "./pages/upload-content";
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" exact element={<Home />} />
+        <Route path="/" element={<Home />} />
         <Route path="/upload-content" element={<UploadContent />} />
       </Routes>
     </BrowserRouter>
