+++ portal-app/src/components/HomePage.jsx
