--- portal-app/src/pages/Home.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import HomeContainer from '../containers/HomeContainer';
-import useUserData from '../hooks/useUserData';
-
-const Home = () => {
-  const { user } = useUserData();
-
-  return (
-    <div>
-      <HomeContainer user={user} />
-    </div>
-  );
-};
-
-export default Home;
